Added configurable JSC threshold, cosmic timing flags for cosmic running, and ability to specify eventSetup labels

+++ L1Trigger/RegionalCaloTrigger/src/L1RCTElectronIsolationCard.cc
@@ -271,7 +271,7 @@ L1RCTElectronIsolationCard::calcElectronCandidates(L1RCTRegion* region, int regi
 	//int quietThreshold = 3;   // 3 - loose isolation 0 - very tight isolation
 	//int quietThreshold = 7; // ECALGREN
 	//int quietThreshold = 0; // HCALGREN
-	double quietThreshold = rctLookupTables_->rctParameters()->eicIsolationThreshold();
+	unsigned quietThreshold = rctLookupTables_->rctParameters()->eicIsolationThreshold();
 	
 	bool nw = false;
 	bool ne = false;
